+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This is the changelog for [AWS Architect](readme.md).
 ## 6.2 ##
 * Allow setting upload zip file directly to support creating lambda layers
 * Add `publishZipArchive` to publish layers and other zip files directly to S3, using the package name and version automatically.
+* New option in `publishLambdaArtifactPromise` to allow turning off running npm or yarn for package deployment `autoHandleCompileOfSourceDirectory = false`.
 
 ## 6.1 ##
 * Deploy CF templates to S3 deployment bucket before deploying to CF to increase allow size of templates to 450KB.