+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change log
 This is the changelog for [AWS Architect](readme.md).
 
+## 6.3 ##
+Automatically delete the stage lambda alias and version when deleting a stage by passing in the lambda function name as the second parameter to `awsArchitect.removeStagePromise`.
+
 ## 6.2 ##
 * Allow setting upload zip file directly to support creating lambda layers
 * Add `publishZipArchive` to publish layers and other zip files directly to S3, using the package name and version automatically.